template change

+++ MinGW/include/debug.h
@@ -1,11 +1,11 @@
 #include<bits/stdc++.h>
 using namespace std;
-typedef int_fast64_t ll;
+typedef long long ll;
 #include<bits/stdtr1c++.h>
 #include<ext/pb_ds/assoc_container.hpp>
 #include<ext/pb_ds/tree_policy.hpp>
 using namespace __gnu_pbds;
-template<class T> 
+template<class T>
 using ordered_set = tree<T, null_type, less_equal<T>, rb_tree_tag, tree_order_statistics_node_update>;
 // #define debug(...) cerr<<"[" << #__VA_ARGS__ <<"] --> "; full_debug(__VA_ARGS__); cerr << endl
 /////////////////////////////
@@ -31,7 +31,7 @@ string __elapsed__(){
     return "";
 }
 void full_debug(int x){cerr << x;}
-void full_debug(int_fast64_t x){cerr << x;}
+void full_debug(long long x){cerr << x;}
 void full_debug(unsigned long long x){cerr << x; }
 void full_debug(char x){cerr << x;}
 void full_debug(string x){cerr << x;}
@@ -42,8 +42,8 @@ void full_debug(long double x){cerr << x;}
 
 // data stractures
 template<class T,class V>
-void full_debug(pair<T,V> x) 
-{ 
+void full_debug(pair<T,V> x)
+{
 	cerr << "{";
 	full_debug(x.first);
 	cerr << ",";
@@ -98,8 +98,8 @@ void full_debug(multiset<T>s){
 	cerr <<"\n]";
 }
 template<class T,class V>
-void full_debug(map<T,V>m) 
-{ 
+void full_debug(map<T,V>m)
+{
 	cerr<<"Map: " << "Size: "<<m.size()<<endl;
 	cerr << "{\n";
 	ll indx = 0;
@@ -115,8 +115,8 @@ void full_debug(map<T,V>m)
 	cerr << "\n}";
 }
 template<class T>
-void full_debug(stack<T>stk) 
-{ 
+void full_debug(stack<T>stk)
+{
 	cerr<<"Stack: " << "Size: "<<stk.size()<<endl;
 	cerr << "[ \n";
 	int i = 0;
@@ -128,7 +128,7 @@ void full_debug(stack<T>stk)
 		else{
 			cerr<<"\t\t ";
 		}
-		
+
 		full_debug(stk.top());cerr <<"\t i: "<<i<<"\t";
 		cerr<<"\t|";
 		cerr <<"\n";
@@ -138,8 +138,8 @@ void full_debug(stack<T>stk)
 	cerr << "\n]";
 }
 template<class T>
-void full_debug(queue<T>Q) 
-{ 
+void full_debug(queue<T>Q)
+{
 	cerr<<"Deque: " << "Size: "<<Q.size()<<endl;
 	cerr << "[ \n";
 	int i = 0;
@@ -151,7 +151,7 @@ void full_debug(queue<T>Q)
 		else{
 			cerr<<"\t\t ";
 		}
-		
+
 		full_debug(Q.front());cerr <<"\t i: "<<i<<"\t";
 		cerr<<"\t|";
 		cerr <<"\n";
@@ -161,12 +161,12 @@ void full_debug(queue<T>Q)
 	cerr << "\n]";
 }
 template<class T>
-void full_debug(deque<T>Q) 
-{ 
+void full_debug(deque<T>Q)
+{
 	cerr<<"Deque: " << "Size: "<<Q.size()<<endl;
 	cerr << "[\n";
 	int i = 0;
-	int n = Q.size(); 
+	int n = Q.size();
 	while(!Q.empty()){
 		cerr<<"\t| ";
 		if(!i){
@@ -178,7 +178,7 @@ void full_debug(deque<T>Q)
 		else{
 			cerr<<"\t\t ";
 		}
-		
+
 		full_debug(Q.front()); cerr <<"\t i: "<<i<<"\t";
 		cerr<<"\t|";
 		cerr <<"\n";
@@ -192,7 +192,7 @@ void full_debug(deque<T>Q)
 
 // premitive types
 void _debug(int x){ cerr << x; }
-void _debug(int_fast64_t x){ cerr << x; }
+void _debug(long long x){ cerr << x; }
 void _debug(unsigned long long x){ cerr << x; }
 void _debug(char x){ cerr << x; }
 void _debug(string x){ cerr << x; }
@@ -213,8 +213,8 @@ void _debug(T x, int mode){
 }
 // data stractures
 template<class T,class V>
-void _debug(pair<T,V> x) 
-{ 
+void _debug(pair<T,V> x)
+{
 	cerr << "{";
 	_debug(x.first);
 	cerr << ",";
@@ -268,8 +268,8 @@ void _debug(ordered_set<T>s){
 }
 
 template<class T,class V>
-void _debug(map<T,V>m) 
-{ 
+void _debug(map<T,V>m)
+{
 	cerr << "{ ";
 	for(auto i : m)
 	{
@@ -282,8 +282,8 @@ void _debug(map<T,V>m)
 	cerr << " }";
 }
 template<class T>
-void _debug(stack<T>stk) 
-{ 
+void _debug(stack<T>stk)
+{
 	cerr << "[ top -> ";
 	while(!stk.empty()){
 		_debug(stk.top());
@@ -293,8 +293,8 @@ void _debug(stack<T>stk)
 	cerr << " ]";
 }
 template<class T>
-void _debug(queue<T>stk) 
-{ 
+void _debug(queue<T>stk)
+{
 	cerr << "[ frnt -> ";
 	while(!stk.empty()){
 		_debug(stk.front());
